+++ app/Resources/doc/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.0.4.3
+* Added keywords to searchfield.
+* Matching all groups in MetadataSearch service.
+* Added some filter columns to metadata entity.
+* Set role string length to 255.
+* Tag metadata as system change befor update public state.
+* Changed BBOX Label from "N O S W" to "W S O N".
+
 ## 2.0.4.2
 * Metadata entity date now nullable.
 * Minor bugfixes in search bundle.